julia_port v0.1.0 JuliaPort

example project to invoke julia functions in elixir to do scientific computing using port and metaprogramming

Functions

*(a0, a1, a2, a3)

helper function defined in GenFunction

complex_test(port)

example to do linear algebra

include_script(port, path)

include a script in julia repl

init()

open a port to start a julia process

init_network(a0, a1, a2)

helper function defined in GenFunction

load_data(a0, a1, a2)

helper function defined in GenFunction

loop(verbose?, data)

helper function to recieve messages

remark: one may modify this function to customise output format

lr_test(a0, a1, a2, a3, a4)

helper function defined in GenFunction

lr_train(a0, a1, a2, a3)

helper function defined in GenFunction

net_eval(a0, a1, a2, a3)

helper function defined in GenFunction

port_receive(port, verbose?)

recieve messages from a port

remark: a trick to use Ω and ω as finished signal

port_send(port, command)

send a command through a port

rand(a0, a1, a2, a3)

helper function defined in GenFunction

real_test(port)

example to do neural network

prerequisite: BackpropNeuralNet installed

script_test(port)

example to do linear regression

simple_test(port)

example to do arithmetics

sum(a0, a1, a2)

helper function defined in GenFunction

terminate(port)

close a port to end a julia process

train(a0, a1, a2, a3, a4)

helper function defined in GenFunction