julia_port v0.2.0 JuliaPort.GenFunction

extensible interface of julia function definition in elixir using macro

Functions

__using__(opts) (macro)

the macro to define functions with dynamic name and arity